requirements.txt
# Core GIS
osmnx
geopandas
shapely
pyproj
rasterio
fiona

# Database
sqlalchemy
geoalchemy2
psycopg2-binary
alembic

# API
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx

# Data Processing
pandas
numpy

# Real-time & Scheduling
requests
apscheduler

# ML / Analytics
scikit-learn
joblib

# Dashboard
streamlit
streamlit-folium
folium
plotly
pydeck

# Utilities
python-dotenv
loguru

# Testing
pytest
pytest-cov

# Code Quality
ruff


pydantic-settings